requirements.txt
fastapi
uvicorn
httpx
pytest
python-dotenv
pytest-asyncio
respx
black
isort
pytest-cov